[Feat] [GGFE-29] fix season types
